+++ rememberall/src/main/java/com/stella/rememberall/dongdong/DongdongImg.java
@@ -1,10 +1,10 @@
-package com.stella.rememberall.domain;
+package com.stella.rememberall.dongdong;
 
 import javax.persistence.*;
 
 @Table(name = "dongdong_img")
 @Entity
-public class DongDongImg {
+public class DongdongImg {
     @Id
     @Column(name = "dongdong_img_id")
     private Long id;
